home_dash_api/pyproject.toml
[tool.aerich]
tortoise_orm = "app.db.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."

[tool.poetry]
name = "home_dash_api"
version = "0.1.0"
description = ""
authors = ["James Best <jamesbest.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
isort = "^5.10.1"
black = "^22.6.0"
flake8 = "^5.0.4"
fastapi = "^0.79.0"
tortoise-orm = "^0.19.2"
asyncpg = "^0.26.0"
aerich = "^0.6.3"
uvicorn = "^0.18.2"
requests = "^2.28.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
PyJWT = "^2.4.0"
python-multipart = "^0.0.5"
autopep8 = "^1.7.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
